assembler and its tests refactored and generalized, immediate and zero page modes added

+++ utilities.h
@@ -2,14 +2,14 @@
 
 #include <cstdint>
 
-constexpr uint8_t loByte(uint16_t word) {
+constexpr uint8_t loByte(unsigned word) {
   return word & 0xff;
 }
 
-constexpr uint8_t hiByte(uint16_t word) {
+constexpr uint8_t hiByte(unsigned word) {
   return word >> 8 & 0xff;
 }
 
-constexpr uint16_t wordOf(uint8_t low, uint16_t high) {
+constexpr uint16_t wordOf(uint8_t low, uint8_t high) {
   return static_cast<uint16_t>(low | high << 8);
 }
